+++ lib/Extension/Completion/CompletionExtension.php
@@ -3,12 +3,13 @@
 namespace Phpactor\Extension\Completion;
 
 use Phpactor\Completion\Bridge\TolerantParser\SourceCodeFilesystem\ScfClassCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseConstantCompletor;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\FunctionFormatter;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\MethodFormatter;
 use Phpactor\Completion\Bridge\WorseReflection\Formatter\VariableFormatter;
 use Phpactor\Completion\Bridge\TolerantParser\ChainTolerantCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseClassMemberCompletor;
-use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseBuiltInFunctionCompletor;
+use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseFunctionCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseParameterCompletor;
 use Phpactor\Completion\Bridge\TolerantParser\WorseReflection\WorseLocalVariableCompletor;
 use Phpactor\Completion\Core\Formatter\ObjectFormatter;
@@ -95,11 +96,15 @@ class CompletionExtension implements Extension
             );
         }, [ 'completion.tolerant_completor' => []]);
 
-        $container->register('completion.completor.builtin_function', function (Container $container) {
-            return new WorseBuiltInFunctionCompletor(
+        $container->register('completion.completor.function', function (Container $container) {
+            return new WorseFunctionCompletor(
                 $container->get('reflection.reflector'),
                 $container->get('completion.formatter')
             );
+        }, [ 'completion.tolerant_completor' => []]);
+
+        $container->register('completion.completor.constant', function (Container $container) {
+            return new WorseConstantCompletor();
         }, [ 'completion.tolerant_completor' => []]);
 
         $container->register('completion.formatter', function (Container $container) {